added test framework and export cpu as a library

+++ main.c
@@ -3,7 +3,6 @@
 # include <stdlib.h>
 
 # include <stdio.h>
-# include "SDL3/SDL.h"
 # include <stdbool.h>
 
 /*
@@ -13,10 +12,9 @@
 static unsigned long master_clock_cycle_count = 0;
 
 int main(){
-
   // fake RAM since its not important now
   uint16_t bus[0xFFFE];
-  for(int i=0;i<sizeof(bus)/sizeof(bus[0]);i++){
+  for(uint16_t i=0;i<sizeof(bus)/sizeof(bus[0]);i++){
     bus[i] = 0x16;
   }
